requirements.txt
torch>=1.9.0
transformers>=4.11.0
pandas>=1.3.0
numpy>=1.19.5
scikit-learn>=0.24.2
tqdm>=4.62.0
faiss-cpu>=1.7.0
langchain>=0.0.200
openai>=0.27.0
anthropic>=0.3.0
python-dotenv>=0.19.0
flask>=2.0.1
requests>=2.26.0
matplotlib>=3.5.0
seaborn>=0.11.0 